+++ src/Uri/InvalidUriException.php
@@ -12,6 +12,8 @@
 namespace Webmozart\Puli\Uri;
 
 /**
+ * Thrown when a URI is invalid.
+ *
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */