+++ src/Uri/UriRepository.php
@@ -11,10 +11,73 @@
 
 namespace Webmozart\Puli\Uri;
 
+use Webmozart\Puli\InvalidPathException;
 use Webmozart\Puli\Resource\Collection\ResourceCollection;
+use Webmozart\Puli\Resource\Collection\ResourceCollectionInterface;
+use Webmozart\Puli\Resource\ResourceInterface;
+use Webmozart\Puli\ResourceNotFoundException;
 use Webmozart\Puli\ResourceRepositoryInterface;
 
 /**
+ * A repository which delegates to other repositories based on URI schemes.
+ *
+ * Repositories can be registered for specific URI schemes. Resource requests
+ * for URIs with that scheme will then be routed to the appropriate
+ * repository:
+ *
+ * ```php
+ * use Webmozart\Puli\ResourceRepository;
+ * use Webmozart\Puli\Uri\UriRepository;
+ *
+ * $puliRepo = new ResourceRepository();
+ * $psr4Repo = new ResourceRepository();
+ *
+ * $repo = new UriRepository();
+ * $repo->register('puli', $puliRepo);
+ * $repo->register('psr4', $psr4Repo);
+ *
+ * $resource = $repo->get('puli:///css/style.css');
+ * // => $puliRepo->get('/css/style.css');
+ *
+ * $resource = $repo->get('psr4:///Webmozart/Puli/Puli.php');
+ * // => $psr4Repo->get('/Webmozart/Puli/Puli.php');
+ * ```
+ *
+ * If not all repositories are needed in every request, you can pass callables
+ * which create the repository on demand:
+ *
+ * ```php
+ * use Webmozart\Puli\ResourceRepository;
+ * use Webmozart\Puli\Uri\UriRepository;
+ *
+ * $repo = new UriRepository();
+ * $repo->register('puli', function () {
+ *     $repo = new ResourceRepository();
+ *     // configuration...
+ *
+ *     return $repo;
+ * });
+ * ```
+ *
+ * The first registered scheme is also registered as default scheme. When
+ * reading paths from that scheme, the protocol may be omitted:
+ *
+ * ```php
+ * use Webmozart\Puli\ResourceRepository;
+ * use Webmozart\Puli\Uri\UriRepository;
+ *
+ * $puliRepo = new ResourceRepository();
+ * $psr4Repo = new ResourceRepository();
+ *
+ * $repo = new UriRepository();
+ * $repo->register('puli', $puliRepo);
+ *
+ * $resource = $repo->get('/css/style.css');
+ * // => $puliRepo->get('/css/style.css');
+ * ```
+ *
+ * The default scheme can be changed with {@link setDefaultScheme}.
+ *
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
@@ -25,10 +88,26 @@ class UriRepository implements UriRepositoryInterface
      */
     private $repos = array();
 
-    public function register($scheme, $repoFactory)
+    /**
+     * @var string|null
+     */
+    private $defaultScheme;
+
+    /**
+     * Registers a repository for a given scheme.
+     *
+     * The repository may either be passed as {@link ResourceRepositoryInterface}
+     * or as callable. If a callable is passed, the callable is invoked as soon
+     * as the scheme is used for the first time. The callable should return a
+     * {@link ResourceRepositoryInterface} object.
+     *
+     * @param string                               $scheme            A URI scheme
+     * @param callable|ResourceRepositoryInterface $repositoryFactory The repository to use
+     */
+    public function register($scheme, $repositoryFactory)
     {
-        if (!$repoFactory instanceof ResourceRepositoryInterface
-                && !is_callable($repoFactory)) {
+        if (!$repositoryFactory instanceof ResourceRepositoryInterface
+                && !is_callable($repositoryFactory)) {
             throw new \InvalidArgumentException(
                 'The locator factory should be a callable or an instance '.
                 'of "Webmozart\Puli\Locator\ResourceLocatorInterface".'
@@ -42,47 +121,164 @@ class UriRepository implements UriRepositoryInterface
             ));
         }
 
-        if (!ctype_alpha($scheme)) {
+        if (!ctype_alnum($scheme)) {
             throw new \InvalidArgumentException(sprintf(
-                'The scheme "%s" should consist of letters only.',
+                'The scheme "%s" should consist of letters and digits only.',
                 $scheme
             ));
         }
 
-        $this->repos[$scheme] = $repoFactory;
+        if (!ctype_alpha($scheme[0])) {
+            throw new \InvalidArgumentException(sprintf(
+                'The first character of the scheme "%s" should be a letter.',
+                $scheme
+            ));
+        }
+
+        $this->repos[$scheme] = $repositoryFactory;
+
+        if (null === $this->defaultScheme) {
+            $this->defaultScheme = $scheme;
+        }
     }
 
+    /**
+     * Unregisters the given scheme.
+     *
+     * Unknown schemes are ignored.
+     *
+     * @param string $scheme A URI scheme.
+     */
     public function unregister($scheme)
     {
         unset($this->repos[$scheme]);
+
+        if ($scheme === $this->defaultScheme) {
+            $this->defaultScheme = null;
+        }
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getSupportedSchemes()
     {
         return array_keys($this->repos);
     }
 
+    /**
+     * {@inheritdoc}
+     */
+    public function getDefaultScheme()
+    {
+        return $this->defaultScheme;
+    }
+
+    /**
+     * Sets the scheme to use when paths are passed instead of URIs.
+     *
+     * When a default scheme is registered, paths are automatically translated
+     * to URIs:
+     *
+     * ```php
+     * $repo->setDefaultScheme('puli');
+     *
+     * $repo->get('/css/style.css')
+     * // => $repo->get('puli:///css/style.css');
+     * ```
+     *
+     * Unless this method is called, the first registered scheme is the default
+     * scheme.
+     *
+     * @param string $scheme A URI scheme.
+     *
+     * @throws UnsupportedSchemeException If the URI scheme is not supported.
+     */
+    public function setDefaultScheme($scheme)
+    {
+        if (!isset($this->repos[$scheme])) {
+            throw new UnsupportedSchemeException(sprintf(
+                'The scheme "%s" was not registered.',
+                $scheme
+            ));
+        }
+
+        $this->defaultScheme = $scheme;
+    }
+
+    /**
+     * Returns the resource at the given URI.
+     *
+     * @param string $uri The URI to the resource. If a path is passed, the
+     *                    default scheme is prepended.
+     *
+     * @return ResourceInterface The resource at this URI.
+     *
+     * @throws ResourceNotFoundException If the resource cannot be found.
+     * @throws InvalidUriException If URI is invalid.
+     * @throws InvalidPathException If the path part of the URI is invalid.
+     * @throws UnsupportedSchemeException If the URI scheme is not supported.
+     */
     public function get($uri)
     {
         $parts = Uri::parse($uri);
 
+        if ('' === $parts['scheme']) {
+            $parts['scheme'] = $this->defaultScheme;
+        }
+
         return $this->getRepository($parts['scheme'])->get($parts['path']);
     }
 
+    /**
+     * Returns the resources matching the given URI.
+     *
+     * @param string $uri A URI that may contain wildcards. If a path is passed,
+     *                    the default scheme is prepended.
+     *
+     * @return ResourceCollectionInterface The resources matching the URI.
+     *
+     * @throws InvalidUriException If URI is invalid.
+     * @throws InvalidPathException If the path part of the URI is invalid.
+     * @throws UnsupportedSchemeException If the URI scheme is not supported.
+     */
     public function find($uri)
     {
         $parts = Uri::parse($uri);
 
+        if ('' === $parts['scheme']) {
+            $parts['scheme'] = $this->defaultScheme;
+        }
+
         return $this->getRepository($parts['scheme'])->find($parts['path']);
     }
 
+    /**
+     * Returns whether any resources match the given URI.
+     *
+     * @param string $uri A URI that may contain wildcards. If a path is passed,
+     *                    the default scheme is prepended.
+     *
+     * @return bool Returns whether any resources exist that match the URI.
+     *
+     * @throws InvalidUriException If URI is invalid.
+     * @throws InvalidPathException If the path part of the URI is invalid.
+     * @throws UnsupportedSchemeException If the URI scheme is not supported.
+     */
     public function contains($uri)
     {
         $parts = Uri::parse($uri);
 
+        if ('' === $parts['scheme']) {
+            $parts['scheme'] = $this->defaultScheme;
+        }
+
         return $this->getRepository($parts['scheme'])->contains($parts['path']);
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getByTag($tag)
     {
         $resources = array();
@@ -96,6 +292,9 @@ class UriRepository implements UriRepositoryInterface
         return new ResourceCollection($resources);
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getTags()
     {
         $tags = array();
@@ -109,10 +308,21 @@ class UriRepository implements UriRepositoryInterface
         return array_keys($tags);
     }
 
+    /**
+     * If necessary constructs and returns the repository for the given scheme.
+     *
+     * @param string $scheme A URI scheme.
+     *
+     * @return ResourceRepositoryInterface The resource repository.
+     *
+     * @throws RepositoryFactoryException If the callable did not return an
+     *                                    instance of {@link ResourceRepositoryInterface}.
+     * @throws UnsupportedSchemeException If the scheme is not supported.
+     */
     private function getRepository($scheme)
     {
         if (!isset($this->repos[$scheme])) {
-            throw new SchemeNotSupportedException(sprintf(
+            throw new UnsupportedSchemeException(sprintf(
                 'The scheme "%s" is not supported.',
                 $scheme
             ));