+++ src/Uri/UriRepositoryInterface.php
@@ -14,13 +14,27 @@ namespace Webmozart\Puli\Uri;
 use Webmozart\Puli\ResourceRepositoryInterface;
 
 /**
+ * A repository which supports URIs for retrieving resources.
+ *
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
 interface UriRepositoryInterface extends ResourceRepositoryInterface
 {
     /**
-     * @return string[]
+     * Returns the supported URI schemes.
+     *
+     * The URI scheme is the part before the "://" in the URL.
+     *
+     * @return string[] The supported URI schemes.
      */
     public function getSupportedSchemes();
+
+    /**
+     * Returns the scheme prepended when a path is passed instead of a URI.
+     *
+     * @return string|null The default scheme or null if no schemes are
+     *                     supported.
+     */
+    public function getDefaultScheme();
 }