+++ src/Util/Selector.php
@@ -12,16 +12,37 @@
 namespace Webmozart\Puli\Util;
 
 /**
+ * Utility methods for handling path selectors.
+ *
+ * "Path selectors" are repository paths which may contain "*" as wildcard.
+ *
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
 class Selector
 {
+    /**
+     * Converts a selector to a regular expression.
+     *
+     * @param string $selector A path selector in canonical form.
+     *
+     * @return string The regular expression for matching the selector.
+     */
     public static function toRegEx($selector)
     {
         return '~^'.str_replace('\*', '[^/]+', preg_quote($selector, '~')).'$~';
     }
 
+    /**
+     * Converts a selector to a glob pattern.
+     *
+     * The flag {@link GLOB_BRACE} must be used if this pattern is passed to
+     * {@link glob}.
+     *
+     * @param string $selector A path selector in canonical form.
+     *
+     * @return string The glob for find files for the selector.
+     */
     public static function toGlob($selector)
     {
         // Return hidden files (starting with ".") when the "*" wildcard is
@@ -29,6 +50,17 @@ class Selector
         return str_replace('/*', '/{.,}*', $selector);
     }
 
+    /**
+     * Returns the static prefix of a selector.
+     *
+     * The "static prefix" are all characters up to the first wildcard "*".
+     * If the selector does not contain wildcards, the full selector is
+     * returned.
+     *
+     * @param string $selector A path selector in canonical form.
+     *
+     * @return string The static prefix of the selector.
+     */
     public static function getStaticPrefix($selector)
     {
         if (false !== ($pos = strpos($selector, '*'))) {